fix: restructure dependency chain
