+++ integration-tests/testkit/flow.ts
@@ -1,6 +1,5 @@
 import { gql } from '@app/gql';
 import { fetch } from '@whatwg-node/fetch';
-import { OperationBodyByHashInput } from './../../packages/web/app/src/graphql/index';
 import type {
   AnswerOrganizationTransferRequestInput,
   CreateOrganizationInput,
@@ -10,6 +9,7 @@ import type {
   DeleteTokensInput,
   EnableExternalSchemaCompositionInput,
   InviteToOrganizationByEmailInput,
+  OperationBodyByHashInput,
   OperationsStatsSelectorInput,
   OrganizationMemberAccessInput,
   OrganizationSelectorInput,
