+++ packages/services/cdn-worker/src/aws.ts
